moove APPS_DETAILS to utils lib
