+++ libs/utils/src/index.ts
@@ -12,3 +12,4 @@ export * from './lib/form/matchers';
 export * from './lib/spreadsheet/export';
 export * from './lib/spreadsheet/import';
 export { OrganizationAlgoliaResult } from './lib/algolia';
+export * from './lib/apps';
